EIP-712 typed data abstraction
Replace ad-hoc field-by-field state with a value-tree
(s_struct_712_value). typed_data.c becomes tree/struct-def storage
and traversal helpers (visitor pattern), exposed under a td_ prefix.
