+++ src/features/provide_gating/cmd_get_gating.c
@@ -480,7 +480,7 @@ static bool check_gating_address(void) {
             selector = GATING->hash_selector;
             break;
         case TX_TYPE_TYPED_DATA:
-            if (!impl_get_domain_contract_addr(eip712_contract)) {
+            if (!td_get_domain_contract_addr(eip712_contract)) {
                 return false;
             }
             contract = eip712_contract;
@@ -545,7 +545,7 @@ static bool check_gating_chain_id(void) {
             break;
         case TX_TYPE_TYPED_DATA:
             // return value not checked on purpose
-            impl_get_domain_chain_id(&chain_id);
+            td_get_domain_chain_id(&chain_id);
             // For EIP-712, the chain_id is optional, and be 0 in the descriptor (any chain)
             if ((GATING->chain_id != 0) && (GATING->chain_id != chain_id)) {
                 PRINTF("[GATING] Chain_ID mismatch: %llu != %llu\n", GATING->chain_id, chain_id);
