+++ src/features/sign_message_eip712/typed_data.h
@@ -71,35 +71,39 @@ typedef struct {
 
 #define TD_MAX_DEPTH 16
 
-const char *get_struct_field_typename(const s_struct_712_field *ptr);
-const s_struct_712 *get_struct_list(void);
-const s_struct_712 *get_structn(const char *name_ptr);
-bool set_struct_name(uint8_t length, const uint8_t *name);
-bool set_struct_field(uint8_t length, const uint8_t *data);
-bool typed_data_init(void);
-void typed_data_deinit(void);
+bool td_add_struct_def(const s_struct_712 *struct_def);
+bool td_add_struct_field_def(s_struct_712 *struct_def, const s_struct_712_field *field_def);
+void td_delete_struct_field(s_struct_712_field *field);
 
-bool impl_set_root(const char *name);
-bool impl_new_array(size_t count);
-// Returns the completed leaf on final chunk, in-progress leaf on intermediate chunks, NULL on
-// error.
-const s_struct_712_value *impl_add_field(const uint8_t *data, size_t length, bool more);
-bool impl_is_complete(void);
-bool impl_hash_pass(void);
+const char *td_get_struct_field_typename(const s_struct_712_field *ptr);
+const s_struct_712 *td_find_struct(const char *name_ptr);
 
-e_root_type impl_get_root_type(void);
-const s_struct_712_field *impl_get_current_field(void);
-uint8_t impl_get_depth_count(void);
-const s_struct_712_field *impl_get_nth_field(uint8_t n);
-uint8_t impl_backup_get_depth_count(void);
-const s_struct_712_field *impl_backup_get_nth_field(uint8_t n);
-bool impl_backup_exists(const char *path, size_t length);
+bool td_set_domain(s_struct_712_value *node);
+bool td_set_message(s_struct_712_value *node);
 
-bool impl_get_domain_chain_id(uint64_t *chain_id);
-bool impl_get_domain_contract_addr(uint8_t addr[ADDRESS_LENGTH]);
+bool td_has_domain(void);
+bool td_has_message(void);
+
+bool td_init(void);
+void td_deinit(void);
+
+bool td_hash_pass(void);
+
+bool td_get_domain_chain_id(uint64_t *chain_id);
+bool td_get_domain_contract_addr(uint8_t addr[ADDRESS_LENGTH]);
 
 // Visitor callback for value tree traversal. Return false to abort traversal.
-typedef bool (*f_value_visitor)(const s_struct_712_value *node, void *context);
+typedef bool (*td_f_value_visitor)(const s_struct_712_value *node, void *context);
 
-bool impl_traverse_domain(f_value_visitor visitor, void *context);
-bool impl_traverse_message(f_value_visitor visitor, void *context);
+bool td_traverse_domain(td_f_value_visitor visitor, void *context);
+bool td_traverse_message(td_f_value_visitor visitor, void *context);
+
+typedef bool (*td_f_struct_visitor)(const s_struct_712 *node, void *context);
+
+bool td_visit_structs(td_f_struct_visitor visitor, void *context);
+
+typedef bool (*td_f_struct_field_visitor)(const s_struct_712_field *node, void *context);
+
+bool td_visit_struct_fields(const s_struct_712 *s,
+                            td_f_struct_field_visitor visitor,
+                            void *context);
